refactor: 문자열 계산식 추출 클래스 분리
